+++ components/timer.js
@@ -1,0 +1,53 @@
+// This code was taken from a metronome tutorial THe github repo was
+//https://github.com/musicandcode/Metronome/blob/main/
+// Add accurate timer constructor function
+
+function Timer(callback, timeInterval, options, isPlaying) {
+    this.timeInterval = timeInterval;
+    
+    // Add method to start timer
+    this.start = () => {
+      // Set the expected time. The moment in time we start the timer plus whatever the time interval is. 
+      this.expected = Date.now() + this.timeInterval;
+      // Start the timeout and save the id in a property, so we can cancel it later
+      this.theTimeout = null;
+      
+      if (options.immediate) {
+        callback();
+      } 
+      
+      this.timeout = setTimeout(this.round, this.timeInterval);
+      console.log('Timer Started');
+    }
+    // Add method to stop timer
+    this.stop = async() => {
+  
+      await clearTimeout(this.timeout);
+      console.log('Timer Stopped');
+    }
+    // Round method that takes care of running the callback and adjusting the time
+    this.round = () => {
+      console.log('timeout', this.timeout);
+      // The drift will be the current moment in time for this round minus the expected time..
+      let drift = Date.now() - this.expected;
+      // Run error callback if drift is greater than time interval, and if the callback is provided
+      if (drift > this.timeInterval) {
+        // If error callback is provided
+        if (options.errorCallback) {
+          options.errorCallback();
+        }
+      }
+      console.log(isPlaying);
+      if (isPlaying==true) {
+        callback();
+      }
+      // Increment expected time by time interval for every round after running the callback function.
+      this.expected += this.timeInterval;
+      console.log('Drift:', drift);
+      console.log('Next round time interval:', this.timeInterval - drift);
+      // Run timeout again and set the timeInterval of the next iteration to the original time interval minus the drift.
+      this.timeout = setTimeout(this.round, this.timeInterval - drift);
+    }
+  }
+
+  export default Timer;
